+++ actionmailer/lib/action_mailer/gem_version.rb
@@ -9,8 +9,8 @@ module ActionMailer
   module VERSION
     MAJOR = 8
     MINOR = 0
-    TINY  = 0
-    PRE   = "1"
+    TINY  = 1
+    PRE   = nil
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")
   end