+++ actiontext/lib/action_text/gem_version.rb
@@ -11,8 +11,8 @@ module ActionText
   module VERSION
     MAJOR = 8
     MINOR = 0
-    TINY  = 0
-    PRE   = "1"
+    TINY  = 1
+    PRE   = nil
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")
   end